+++ src/lib/gui/dialogs/SettingsDialog.cpp
