+++ src/lib/gui/dialogs/SettingsDialog.h
