+++ src/lib/gui/dialogs/SettingsDialogBase.ui
